requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-allauth==0.57.0
dj-rest-auth==5.0.2
python-decouple==3.8
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
openai==1.3.5
requests==2.31.0
Pillow==10.1.0
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
